# docs/api/openapi-django.yaml
openapi: 3.0.3
info:
  title: MSME Marketplace API
  version: 1.0.0
  description: 'Django REST API: auth, catalog, orders/bids, disputes, notifications,
    pricing.'
paths:
...
  /api/auth/register/:
    post:
      operationId: auth_register_create
      description: POST /api/auth/register/
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
          description: ''
...
  /api/notifications/{id}/read/:
    post:
      operationId: notifications_read_create
      description: POST /api/notifications/{id}/read/
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - notifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
          description: ''
...
  /api/notifications/read-all/:
    post:
      operationId: notifications_read_all_create
      description: POST /api/notifications/read-all/
      tags:
      - notifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkedRead'
          description: ''
...
components:
  schemas:
...
    MarkedRead:
      type: object
      properties:
        marked_read:
          type: integer
      required:
      - marked_read
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        user:
          type: integer
          readOnly: true
        type:
          $ref: '#/components/schemas/NotificationTypeEnum'
        title:
          type: string
          maxLength: 200
        message:
          type: string
        related_object_type:
          type: string
          maxLength: 50
        related_object_id:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
        is_read:
          type: boolean
        fcm_sent:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - fcm_sent
      - id
      - title
      - type
      - user
    NotificationTypeEnum:
      enum:
      - GRADING_COMPLETE
      - ORDER_MATCHED
      - BID_RECEIVED
      - DISPUTE_UPDATE
      - SYSTEM
      type: string
      description: |-
        * `GRADING_COMPLETE` - Grading Complete
        * `ORDER_MATCHED` - Order Matched
        * `BID_RECEIVED` - Bid Received
        * `DISPUTE_UPDATE` - Dispute Update
        * `SYSTEM` - System
...
    Register:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          maxLength: 20
        role:
          $ref: '#/components/schemas/RoleEnum'
      required:
      - email
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        phone:
          type: string
          maxLength: 20
        password:
          type: string
          writeOnly: true
          minLength: 8
        role:
          $ref: '#/components/schemas/RoleEnum'
        display_name:
          type: string
          writeOnly: true
      required:
      - email
      - password
...
    RoleEnum:
      enum:
      - SELLER
      - BUYER
      - ADMIN
      - VERIFIER
      type: string
      description: |-
        * `SELLER` - Seller
        * `BUYER` - Buyer
        * `ADMIN` - Admin
        * `VERIFIER` - Verifier
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT